algoـresponder/requirements.txt
openai
python-telegram-bot
python-dotenv
motor
pymongo